バリア展開完成 実用性はまだない
