+++ OculusGame/Assets/Resources/Prefabs/Play/Beacon/Babel/babel.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a62744937d421c94cbe45fe33ba412e0
+guid: 2d34d695f16131149880d5f6509d5e75
 timeCreated: 1447979149
 licenseType: Free
 NativeFormatImporter:
