--- OculusGame/Assets/Resources/Prefabs/Play/Beacon/BabelBeacon.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 15972a1eb7e596e44a160eb9de1f4077
-timeCreated: 1448845693
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
