+++ OculusGame/Assets/Resources/Prefabs/Play/barrier/barrier.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 9672685a165b0d74190153a16645a4b8
-timeCreated: 1448431029
+guid: 479ceecebb292544aa6e90c701f002a1
+timeCreated: 1449035881
 licenseType: Free
 NativeFormatImporter:
   userData:
