+++ OculusGame/Assets/Scripts/Barrier/barriercollision.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: dd24d0503f229d64d8227666a0da947f
+timeCreated: 1449032772
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
