+++ OculusGame/Assets/Scripts/Beacon/BeaconUI.cs
@@ -9,9 +9,6 @@ public class BeaconUI : MonoBehaviour
     Slider ber;
     Image clearBer;
     float time;
-    enum barrierState { OPEN,DELETE,STAY,IDLE }; //　barrierステータス、展開中、破壊、待機中
-    barrierState barrierFlg;
-    GameObject barrier;
 
     void Awake()
     {
@@ -21,7 +18,6 @@ public class BeaconUI : MonoBehaviour
         clearBer = GameObject.Find("Background").GetComponent<Image>();
         time = 0;
         clearBer.enabled = false;
-        barrierFlg = barrierState.IDLE;
     }
 
     public void BeaconPutUI()
@@ -31,16 +27,12 @@ public class BeaconUI : MonoBehaviour
             clearBer.enabled = true;
             time += Time.deltaTime;
             ber.value += Time.deltaTime;
-            if (barrierFlg == barrierState.IDLE) {
-                barrierFlg = barrierState.OPEN;
-            }
 
             if (time >= playerBeacon.waitTime)   //時間を超えたとき設置
             {
                 clearBer.enabled = false;
                 ber.value = 0;
                 time = 0;
-                barrierFlg = barrierState.DELETE;
             }
         }
 
@@ -49,28 +41,7 @@ public class BeaconUI : MonoBehaviour
             clearBer.enabled = false;
             ber.value = 0;
             time = 0;
-            barrierFlg = barrierState.DELETE;
         }
 
-    }
-
-    public void BarrierOpen()
-    {
-        switch (barrierFlg)
-        {
-            case barrierState.DELETE:
-                if (GameObject.Find("barrier") != null)
-                    Destroy(barrier);
-                break;
-            case barrierState.OPEN:
-                Instantiate(ResourcesManager.Instance.GetResourceScene("barrier"), transform.position, transform.rotation);
-                barrierFlg = barrierState.STAY;
-                break;
-            case barrierState.STAY:
-                
-                break;
-            default:
-                break;
-        }
     }
 }
